refactor(getNumberOfRounds): 경주 회차 입력 로직을 모듈로 분리
경주 회차 수를 입력받는 메소드 getNumberOfRounds를 App이 아닌 다른 클래스의 메소드로 분리 : Game/UserInputRounds.js

테스트 실황
Test Suites: 2 passed, 2 total
Tests:       7 passed, 7 total
Snapshots:   0 total
Time:        1.049 s
