+++ src/Game/UserInputRounds.js
@@ -1,0 +1,15 @@
+import { MissionUtils } from '@woowacourse/mission-utils';
+import { MESSAGE, ERROR } from '../constants/index.js';
+
+export class UserInputRounds {
+  static async getNumberOfRounds() {
+    const COUNTS = await MissionUtils.Console.readLineAsync(
+      MESSAGE.raceRoundForStart,
+    );
+    const ROUNDS = parseInt(COUNTS, 10);
+    if (isNaN(ROUNDS)) {
+      throw new Error(ERROR.roundInputNumber);
+    }
+    return ROUNDS;
+  }
+}
